requirements.txt
openai
uvicorn[standard]
fastapi
pydantic
pyyaml
chromadb
tiktoken
